various updates
Updated the textfield on the map view to use groups instead of an
enterFrame listener to keep the field in position with its background
chrome.
Added a common font for all devices since HelveticaNeue wasn't
available on some platforms.
Added back buttons to the registration and edituser scenes.
Adjusted content positioning in some places to work better with devices
with status bars.